model criated

+++ src/main/java/com/vendas/models/Produto.java
@@ -1,5 +1,7 @@
 package com.vendas.models;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -11,8 +13,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class Produto {
 
-
-    private int id_produto;
-    private int preco_unitario;
+    private Integer id_produto;
+    private BigDecimal preco_unitario;
     private String descricao;
 }
